# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Workcraft API
  description: |
    Studio-ops API for Workcraft. Response envelope: `{ status: "success"|"error", data? | message? }`.
    Authenticated routes expect `Authorization: Bearer <JWT>`. Domain routes require studio membership on the token.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiSuccess:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: success
        data:
          description: Endpoint-specific payload
...
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
paths:
...
  /api/settings/packages:
    get:
      tags: [Settings]
      summary: List packages
      operationId: listPackages
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Packages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    post:
      tags: [Settings]
      summary: Create package
      operationId: createPackage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                price:
                  oneOf:
                    - type: number
                    - type: string
                description:
                  type: string
                  nullable: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/packages/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Settings]
      summary: Get package
      operationId: getPackage
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    patch:
      tags: [Settings]
      summary: Update package
      operationId: updatePackage
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    delete:
      tags: [Settings]
      summary: Delete package
      operationId: deletePackage
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/payment-plans:
    get:
      tags: [Settings]
      summary: List payment plans
      operationId: listPaymentPlans
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Plans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    post:
      tags: [Settings]
      summary: Create payment plan
      operationId: createPaymentPlan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                milestones:
                  type: array
                  items:
                    type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/payment-plans/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Settings]
      summary: Get payment plan
      operationId: getPaymentPlan
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    patch:
      tags: [Settings]
      summary: Update payment plan (milestones replace when sent)
      operationId: updatePaymentPlan
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    delete:
      tags: [Settings]
      summary: Delete payment plan
      operationId: deletePaymentPlan
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/contracts:
    get:
      tags: [Settings]
      summary: List contract templates
      operationId: listContracts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Contracts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    post:
      tags: [Settings]
      summary: Create contract template
      operationId: createContract
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, body]
              properties:
                name:
                  type: string
                body:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/contracts/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Settings]
      summary: Get contract template
      operationId: getContract
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    patch:
      tags: [Settings]
      summary: Update contract template
      operationId: updateContract
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    delete:
      tags: [Settings]
      summary: Delete contract template
      operationId: deleteContract
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/checklists:
    get:
      tags: [Settings]
      summary: List checklist templates
      operationId: listChecklists
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Checklists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    post:
      tags: [Settings]
      summary: Create checklist template
      operationId: createChecklist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                isDefault:
                  type: boolean
                items:
                  type: array
                  items:
                    type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/checklists/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Settings]
      summary: Get checklist template
      operationId: getChecklist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Checklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    patch:
      tags: [Settings]
      summary: Update checklist template
      operationId: updateChecklist
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    delete:
      tags: [Settings]
      summary: Delete checklist template
      operationId: deleteChecklist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/checklists/{id}/duplicate:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    post:
      tags: [Settings]
      summary: Duplicate checklist template
      operationId: duplicateChecklist
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/email-templates:
    get:
      tags: [Settings]
      summary: List email templates
      operationId: listEmailTemplates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/email-templates/seed:
    post:
      tags: [Settings]
      summary: Seed default email templates
      operationId: seedEmailTemplates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Seeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/email-templates/{type}:
    parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum: [QUOTE, INVOICE, REM_BEFORE, REM_DUE, REM_AFTER]
    put:
      tags: [Settings]
      summary: Upsert email template by type
      operationId: upsertEmailTemplate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, subject, body]
              properties:
                title:
                  type: string
                subject:
                  type: string
                body:
                  type: string
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/whatsapp-templates:
    get:
      tags: [Settings]
      summary: List WhatsApp templates
      operationId: listWhatsappTemplates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/whatsapp-templates/seed:
    post:
      tags: [Settings]
      summary: Seed default WhatsApp templates
      operationId: seedWhatsappTemplates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Seeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/whatsapp-templates/{type}:
    parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum: [QUOTE, INVOICE, REM_BEFORE, REM_DUE, REM_AFTER, CONTRACT]
    put:
      tags: [Settings]
      summary: Upsert WhatsApp template by type
      operationId: upsertWhatsappTemplate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title:
                  type: string
                body:
                  type: string
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
...